+++ test/inductor/test_fp8.py
@@ -13,7 +13,7 @@ from torch.testing._internal.common_utils import (
     parametrize,
     TEST_WITH_ROCM,
 )
-from torch.testing._internal.inductor_utils import HAS_CPU, HAS_CUDA
+from torch.testing._internal.inductor_utils import HAS_CUDA
 from torch.utils._triton import has_triton_tma_device
 
 
@@ -107,10 +107,10 @@ def _fix_fp8_dtype_for_hip(dtype, device):
 
 @instantiate_parametrized_tests
 class TestFP8Types(TestCase):
+    @unittest.skipIf(not PLATFORM_SUPPORTS_FP8, f8_msg)
     @unittest.skipIf(TEST_WITH_ROCM, "Not supported yet")
     @parametrize("float8_dtype", (torch.float8_e4m3fn, torch.float8_e5m2))
-    @parametrize("device", ("cuda", "cpu"))
-    def test_xblock_for_small_numel(self, float8_dtype: torch.dtype, device: str):
+    def test_xblock_for_small_numel(self, float8_dtype: torch.dtype):
         """
         TritonOverrides.to_dtype will set min_elem_per_thread to 2 or 4
         depends on the variant of fp8 type.
@@ -119,23 +119,19 @@ class TestFP8Types(TestCase):
 
         We should not pick a XBLOCK larger than xnumel
         """
-        if device == "cuda" and not PLATFORM_SUPPORTS_FP8:
-            raise unittest.SkipTest(f8_msg)
 
         def f(x):
             return x.to(dtype=float8_dtype)
 
-        x = torch.randn(1, device=device)
+        x = torch.randn(1, device="cuda")
         expected = f(x)
         actual = torch.compile(f)(x)
         torch.testing.assert_close(expected.half(), actual.half(), rtol=1e-2, atol=1e-2)
 
+    @unittest.skipIf(not PLATFORM_SUPPORTS_FP8, f8_msg)
     @unittest.skipIf(TEST_WITH_ROCM, "Not supported yet")
     @parametrize("dtype", (torch.float16, torch.bfloat16))
-    @parametrize("device", ("cuda", "cpu"))
-    def test_eager_fallback(self, dtype: torch.dtype, device: torch.device):
-        if device == "cuda" and not PLATFORM_SUPPORTS_FP8:
-            raise unittest.SkipTest(f8_msg)
+    def test_eager_fallback(self, dtype: torch.dtype):
         weight_shape = (32, 16)
 
         e4m3_type = (
@@ -143,11 +139,11 @@ class TestFP8Types(TestCase):
         )
 
         def fp8_matmul_unwrapped(x):
-            a_scale = torch.Tensor([1.0]).to(device=device)
-            b_scale = torch.Tensor([1.0]).to(device=device)
+            a_scale = torch.Tensor([1.0]).to(device="cuda")
+            b_scale = torch.Tensor([1.0]).to(device="cuda")
             output_scale = None
-            input_bias = torch.rand(32, device=device, dtype=dtype)
-            weight = torch.rand(*weight_shape, device=device, dtype=dtype).T.to(
+            input_bias = torch.rand(32, device="cuda", dtype=dtype)
+            weight = torch.rand(*weight_shape, device="cuda", dtype=dtype).T.to(
                 e4m3_type
             )
             a_inverse_scale = 1 / a_scale
@@ -168,24 +164,21 @@ class TestFP8Types(TestCase):
         )
 
         x_shape = (16, 16)
-        x = torch.rand(*x_shape, device=device, dtype=dtype).to(e4m3_type)
+        x = torch.rand(*x_shape, device="cuda", dtype=dtype).to(e4m3_type)
         y_fp8 = compiled_fp8_matmul(x)  # noqa: F841
 
         x_shape = (15, 16)
-        x = torch.rand(*x_shape, device=device, dtype=dtype).to(e4m3_type)
+        x = torch.rand(*x_shape, device="cuda", dtype=dtype).to(e4m3_type)
         y_fp8 = compiled_fp8_matmul(x)  # noqa: F841
 
+    @unittest.skipIf(not PLATFORM_SUPPORTS_FP8, f8_msg)
     @parametrize("dtype", (torch.float16, torch.bfloat16, torch.float))
     @parametrize("shape", ("15,3,13", "4,2048,4096"))
     @parametrize("dst_types", [(torch.float8_e4m3fn, torch.float8_e5m2)])
-    @parametrize("device", ("cuda", "cpu"))
-    def test_valid_cast(
-        self, dtype: torch.dtype, shape: str, dst_types: tuple, device: torch.device
-    ):
-        if device == "cuda" and not PLATFORM_SUPPORTS_FP8:
-            raise unittest.SkipTest(f8_msg)
-
-        dst_types = _fix_fp8_dtype_for_hip(dst_types, device)
+    def test_valid_cast(self, dtype: torch.dtype, shape: str, dst_types: tuple):
+        
+        dst_types = _fix_fp8_dtype_for_hip(dst_types, device="cuda")
+        
         e4m3, e5m2 = dst_types
 
         def fp8_cast(x):
@@ -196,7 +189,7 @@ class TestFP8Types(TestCase):
         compiled_fp8_cast = torch.compile(fp8_cast, backend="inductor", dynamic=True)
 
         shape = [int(dim) for dim in shape.split(",")]
-        x = torch.rand(*shape, device=device, dtype=dtype)
+        x = torch.rand(*shape, device="cuda", dtype=dtype)
         y0_fp8, y1_fp8 = compiled_fp8_cast(x)
 
         torch.testing.assert_close(y0_fp8, x, rtol=5e-1, atol=5e-1)
@@ -225,21 +218,14 @@ class TestFP8Types(TestCase):
             x = torch.rand(*x_shape, device="cuda").to(dtype=torch.float8_e5m2)
             compiled_fp8_cast(x, torch.float8_e4m3fn)
 
+    @unittest.skipIf(not PLATFORM_SUPPORTS_FP8, f8_msg)
     @parametrize("src_dtype", (torch.float16, torch.bfloat16, torch.float))
     @parametrize("dst_dtype", (torch.float8_e4m3fn, torch.float8_e5m2))
     @parametrize("shape", ("16,16,16", "4,2048,4096"))
-    @parametrize("device", ("cuda", "cpu"))
     def test_to_fp8_saturated(
-        self,
-        src_dtype: torch.dtype,
-        dst_dtype: torch.dtype,
-        shape: str,
-        device: torch.device,
+        self, src_dtype: torch.dtype, dst_dtype: torch.dtype, shape: str
     ):
-        if device == "cuda" and not PLATFORM_SUPPORTS_FP8:
-            raise unittest.SkipTest(f8_msg)
-
-        dst_dtype = _fix_fp8_dtype_for_hip(device)
+        dst_dtype = _fix_fp8_dtype_for_hip(dst_dtype, device="cuda")
 
         def fp8_saturated(x, dtype):
             return _to_fp8_saturated(x, dtype)
@@ -248,13 +234,14 @@ class TestFP8Types(TestCase):
             fp8_saturated, backend="inductor", dynamic=True
         )
         shape = [int(dim) for dim in shape.split(",")]
-        x = torch.rand(*shape, device=device, dtype=src_dtype)
+        x = torch.rand(*shape, device="cuda", dtype=src_dtype)
         y_compiled = compiled_fp8_cast(x, dst_dtype)
         y = fp8_saturated(x, dst_dtype)
 
         torch.testing.assert_close(y_compiled.half(), y.half(), rtol=5e-1, atol=5e-1)
 
     @unittest.skipIf(TEST_WITH_ROCM, "ROCm fails with accuracy issue")
+    @unittest.skipIf(not SM90OrLater, "FP8 is only supported on H100+")
     @parametrize(
         "float8_dtype",
         (
@@ -264,12 +251,7 @@ class TestFP8Types(TestCase):
         ),
     )
     @parametrize("shape", ("1,1,15", "1,10,15", "1,10,512", "1,10,4096", "4,2048,4096"))
-    @parametrize("device", ("cuda", "cpu"))
-    def test_amax_fp8_quant(
-        self, float8_dtype: torch.dtype, shape: str, device: torch.device
-    ):
-        if device == "cuda" and not SM90OrLater:
-            raise unittest.SkipTest("FP8 is only supported on H100+")
+    def test_amax_fp8_quant(self, float8_dtype: torch.dtype, shape: str):
         shape = [int(dim) for dim in shape.split(",")]
         batch_size, sequence_length, hidden_size = shape
 
@@ -282,25 +264,21 @@ class TestFP8Types(TestCase):
         compiled_amax_fp8_quant = torch.compile(amax_fp8, backend="inductor")
 
         x_shape = (batch_size, sequence_length, hidden_size)
-        x = torch.rand(*x_shape, device=device, dtype=torch.half)
-        scale = torch.tensor(0.2, device=device, dtype=torch.float)
+        x = torch.rand(*x_shape, device="cuda", dtype=torch.half)
+        scale = torch.tensor(0.2, device="cuda", dtype=torch.float)
 
         y_compiled = compiled_amax_fp8_quant(x, scale)
         y = amax_fp8(x, scale)
 
         torch.testing.assert_close(y_compiled.half(), y.half(), rtol=1e-2, atol=1e-2)
 
+    @unittest.skipIf(not PLATFORM_SUPPORTS_FP8, f8_msg)
     @parametrize("float8_dtype", (torch.float8_e4m3fn, torch.float8_e5m2))
     @parametrize("shape", ("1,1,15", "1,10,15", "1,10,512", "1,10,4096", "4,2048,4096"))
-    @parametrize("device", ("cuda", "cpu"))
-    def test_amax_along_with_fp8_quant(
-        self, float8_dtype: torch.dtype, shape: str, device: torch.device
-    ):
-        if device == "cuda" and not PLATFORM_SUPPORTS_FP8:
-            raise unittest.SkipTest(f8_msg)
-
-        float8_dtype = _fix_fp8_dtype_for_hip(float8_dtype, device)
-
+    def test_amax_along_with_fp8_quant(self, float8_dtype: torch.dtype, shape: str):
+        
+        float8_dtype = _fix_fp8_dtype_for_hip(float8_dtype, device="cuda")
+        
         shape = [int(dim) for dim in shape.split(",")]
         batch_size, sequence_length, hidden_size = shape
 
@@ -313,12 +291,12 @@ class TestFP8Types(TestCase):
         compiled_amax_fp8_quant = torch.compile(amax_fp8, backend="inductor")
 
         x_shape = (batch_size, sequence_length, hidden_size)
-        x = torch.rand(*x_shape, device=device, dtype=torch.half)
-        scale = torch.tensor(1.0, device=device, dtype=torch.float)
+        x = torch.rand(*x_shape, device="cuda", dtype=torch.half)
+        scale = torch.tensor(1.0, device="cuda", dtype=torch.float)
 
-        amax_buffer_compiled = torch.zeros((1), device=device, dtype=torch.half)
+        amax_buffer_compiled = torch.zeros((1), device="cuda", dtype=torch.half)
         y_compiled = compiled_amax_fp8_quant(x, scale, amax_buffer_compiled)
-        amax_buffer = torch.zeros((1), device=device, dtype=torch.half)
+        amax_buffer = torch.zeros((1), device="cuda", dtype=torch.half)
         y = amax_fp8(x, scale, amax_buffer)
 
         torch.testing.assert_close(y_compiled.half(), y.half(), rtol=1e-1, atol=1e-1)
@@ -327,6 +305,7 @@ class TestFP8Types(TestCase):
         )
 
     @unittest.skipIf(TEST_WITH_ROCM, "ROCm fails with accuracy issue")
+    @unittest.skipIf(not SM90OrLater, "FP8 is only supported on H100+")
     @parametrize(
         "float8_dtype",
         (
@@ -337,16 +316,9 @@ class TestFP8Types(TestCase):
     )
     @parametrize("amax_keep_dim", (True, False))
     @parametrize("shape", ("1,1,15", "1,10,15", "1,10,512", "1,10,4096", "4,2048,4096"))
-    @parametrize("device", ("cuda", "cpu"))
     def test_layernorm_fp8_quant(
-        self,
-        float8_dtype: torch.dtype,
-        amax_keep_dim: bool,
-        shape: str,
-        device: torch.device,
+        self, float8_dtype: torch.dtype, amax_keep_dim: bool, shape: str
     ):
-        if device == "cuda" and not SM90OrLater:
-            raise unittest.SkipTest("FP8 is only supported on H100+")
         shape = [int(dim) for dim in shape.split(",")]
         batch_size, sequence_length, hidden_size = shape
 
@@ -368,12 +340,12 @@ class TestFP8Types(TestCase):
         compiled_ln_fp8_quant = torch.compile(ln_fp8, backend="inductor")
 
         x_shape = (batch_size, sequence_length, hidden_size)
-        x = torch.rand(*x_shape, device=device, dtype=torch.half)
-        scale = torch.tensor(0.2, device=device, dtype=torch.float)
+        x = torch.rand(*x_shape, device="cuda", dtype=torch.half)
+        scale = torch.tensor(0.2, device="cuda", dtype=torch.float)
 
-        amax_buffer_compiled = torch.zeros((1), device=device, dtype=torch.half)
+        amax_buffer_compiled = torch.zeros((1), device="cuda", dtype=torch.half)
         y_compiled = compiled_ln_fp8_quant(x, scale, amax_buffer_compiled)
-        amax_buffer = torch.zeros((1), device=device, dtype=torch.half)
+        amax_buffer = torch.zeros((1), device="cuda", dtype=torch.half)
         y = ln_fp8(x, scale, amax_buffer)
 
         torch.testing.assert_close(y_compiled.half(), y.half(), rtol=1e-1, atol=1e-1)
@@ -793,5 +765,5 @@ class TestFP8Lowering(TestCase):
 
 
 if __name__ == "__main__":
-    if HAS_CUDA or HAS_CPU:
+    if HAS_CUDA:
         run_tests()
