+++ torch/testing/_internal/common_device_type.py
@@ -1008,8 +1008,6 @@ ANY_DTYPE_ORDER = (
     torch.int8,
     torch.uint8,
     torch.bool,
-    torch.float8_e4m3fn,
-    torch.float8_e5m2,
 )
